+++ shub/apps/library/urls.py
@@ -31,19 +31,21 @@ urlpatterns = [
         views.CompletePushImageFileView.as_view(),
     ),
     url(
-        r"^v2/imagefile/(?P<container_id>.+?)/_multipart?$",
+        r"^v2/imagefile/(?P<upload_id>.+?)/_multipart?$",
         views.RequestMultiPartPushImageFileView.as_view(),
-    ),  # added to scs-library-client Feb 2020, needs to
-    # return 404 so default to old push endpoint
+    ),  # added to scs-library-client Feb 2020, returns 404 if multipart is disabled, defaults to old push endpoint
+    url(
+        r"^v2/imagefile/(?P<upload_id>.+?)/_multipart_abort?$",
+        views.RequestMultiPartAbortView.as_view(),
+    ),
+    url(
+        r"^v2/imagefile/(?P<upload_id>.+?)/_multipart_complete?$",
+        views.RequestMultiPartCompleteView.as_view(),
+    ),
     url(
         r"^v2/imagefile/(?P<container_id>.+?)/?$",
         views.RequestPushImageFileView.as_view(),
     ),  # return push url
-    url(
-        r"^v2/push/imagefile/(?P<container_id>.+?)/(?P<secret>.+?)?$",
-        views.PushImageFileView.as_view(),
-        name="PushImageFileView",
-    ),  # push image
     url(r"^v1/token-status$", views.TokenStatusView.as_view()),
     url(
         r"^v1/collections/(?P<username>.+?)/(?P<name>.+?)$",